drm: Add ASPEED GFX driver
This driver is for the ASPEED BMC SoC's GFX display hardware. This
driver runs on the ARM based BMC systems, unlike the ast driver which
runs on a host CPU and is is for a PCI graphics device.

The AST2500 supports a total of 3 output paths:

  1. VGA output, the output target can choose either or both to the DAC
  or DVO interface.

  2. Graphics CRT output, the output target can choose either or both to
  the DAC or DVO interface.

  3. Video input from DVO, the video input can be used for video engine
  capture or DAC display output.

Output options are selected in SCU2C.

The "VGA mode" device is the PCI attached controller. The "Graphics CRT"
is the ARM's internal display controller.

The driver only supports a simple configuration consisting of a 40MHz
pixel clock, fixed by hardware limitations, and the VGA output path.

OpenBMC-Staging-Count: 1
Signed-off-by: Joel Stanley <joel@jms.id.au>