ARM: dts: aspeed: Enable the GFX IP
The GFX controller is the internal graphics device used by the SoC
(opposed to the one connected via the PCIe device and used by the host).

This configures it with a framebuffer region and adds it to the command
line so kernel boot messages appear on the display.

Enabled for Romulus, Witherspoon, and the ASPEED AST2500 EVB.

OpenBMC-Staging-Count: 4
Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>