fsi/fsi-master-gpio: Sample input data on different clock phase
We currently sample the input data right after we toggle the
clock low, then high. The slave establishes the data on the
rising edge, so this is not ideal. We should sample it on
the low phase instead.

This currently works because we have an extra delay, but subsequent
patches will remove it.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Reviewed-by: Christopher Bostic <cbostic@linux.vnet.ibm.com>
Tested-by: Joel Stanley <joel@jms.id.au>
(cherry picked from commit 5d0d16f135c0cb7a12e417cbe12b271524ce4d40)
Signed-off-by: Joel Stanley <joel@jms.id.au>