arm: dts: aspeed: Change how power9 chips are represented on OpenPower
On OpenPower systems, we currently have a single ibm-power9-cfam.dtsi
which represents in fact 2 cascaded power9 chips, one directly off
the FSI bus hanging off the BMC and one cascaded off the FSI master
of that first P9.

This is messy with a lot of duplication and won't work when we
start having more chips.

We replace that with instead a single chip in that file, and a
macro system to generate unique labels for various nodes that
need to be referenced.

We also create a ibm-power9-dual.dtsi "helper" that represents
the standard 2 chips configuration to avoid duplication.

This also adds the new chip-id property defined for FSI slaves.

OpenBMC-Staging-Count: 1
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Joel Stanley <joel@jms.id.au>